.NET Sample migrated to .NET
