--- QueueStorage/packages.config
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Azure.ConfigurationManager" version="4.0.0" targetFramework="net472" />
-  <package id="Microsoft.Azure.KeyVault.Core" version="3.0.3" targetFramework="net472" />
-  <package id="Microsoft.Azure.Storage.Common" version="10.0.2" targetFramework="net472" />
-  <package id="Microsoft.Azure.Storage.Queue" version="10.0.2" targetFramework="net472" />
-  <package id="Microsoft.Data.Edm" version="5.8.4" targetFramework="net472" />
-  <package id="Microsoft.Data.OData" version="5.8.4" targetFramework="net472" />
-  <package id="Microsoft.Data.Services.Client" version="5.8.4" targetFramework="net472" />
-  <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net472" />
-  <package id="System.Spatial" version="5.8.4" targetFramework="net472" />
-</packages>
