+++ README.md
@@ -33,7 +33,7 @@ To run the sample using the Storage Emulator (default option):
 
 To run the sample using the Storage Service
 
-1. Open the app.config file and comment out the connection string for the emulator (UseDevelopmentStorage=True) and uncomment the connection string for the storage service (AccountName=[]...)
+1. Open the appsettings.json file and comment out the connection string for the emulator (UseDevelopmentStorage=True) and uncomment the connection string for the storage service (AccountName=[]...)
 2. Create a Storage Account through the Azure Portal and provide your [AccountName] and [AccountKey] in the App.Config file.
 3. Set breakpoints and run the project using F10.
