+++ appsettings.json
@@ -1,0 +1,3 @@
+{
+  "ConnectionString": "UseDevelopmentStorage=true;"
+}
